Update strings, removed unused
