--- WMFComponents/Sources/WMFComponents/Assets.xcassets/close-circle-inverse.imageset/Contents.json
@@ -1,15 +1,0 @@
-{
-  "images" : [
-    {
-      "filename" : "Close Button.svg",
-      "idiom" : "universal"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  },
-  "properties" : {
-    "preserves-vector-representation" : true
-  }
-}
